+++ src/coupon/application/use-cases/tier-1-in-domain/cancel-user-coupon.use-case.ts
@@ -1,8 +1,10 @@
 import { Injectable } from "@nestjs/common";
 import { UserCoupon } from "@/coupon/domain/entities/user-coupon.entity";
 import { Coupon } from "@/coupon/domain/entities/coupon.entity";
-import { UserCouponNotFoundError } from "@/coupon/domain/exceptions/user-coupon.exception";
-import { CouponNotFoundError } from "@/coupon/domain/exceptions/coupon.exceptions";
+import {
+  UserCouponNotFoundError,
+  CouponNotFoundError,
+} from "@/coupon/application/coupon.application.exceptions";
 import { Transactional } from "typeorm-transactional";
 import { CouponRepository } from "@/coupon/infrastructure/persistence/coupon.repository";
 import { UserCouponRepository } from "@/coupon/infrastructure/persistence/user-coupon.repository";
