+++ src/order/application/use-cases/tier-1-in-domain/change-order-status.use-case.ts
@@ -1,6 +1,6 @@
 import { Inject, Injectable } from "@nestjs/common";
 import { Order, OrderStatus } from "@/order/domain/entities/order.entitiy";
-import { OrderNotFoundError } from "@/order/domain/exceptions/order.exceptions";
+import { OrderNotFoundError } from "@/order/application/order.application.exceptions";
 import { OrderRepository } from "@/order/infrastructure/persistence/order.repository";
 
 export interface ChangeOrderStatusUseCaseCommand {
