+++ src/product/application/use-cases/tier-1-in-domain/confirm-stock.use-case.spec.ts
@@ -1,10 +1,10 @@
 import { Test } from "@nestjs/testing";
 import { ConfirmStockUseCase } from "./confirm-stock.use-case";
 import {
-  StockReservationNotFoundError,
   StockReservationNotActiveError,
   StockReservationExpiredError,
 } from "@/product/domain/exceptions/product.exceptions";
+import { StockReservationNotFoundError } from "@/product/application/product.application.exceptions";
 import { StockReservation } from "@/product/domain/entities/stock-reservation.entity";
 import { Product } from "@/product/domain/entities/product.entity";
 import { v4 as uuidv4 } from "uuid";
