+++ src/product/application/use-cases/tier-1-in-domain/release-stock.use-case.spec.ts
@@ -1,9 +1,7 @@
 import { Test } from "@nestjs/testing";
 import { ReleaseStockUseCase } from "./release-stock.use-case";
-import {
-  StockReservationNotFoundError,
-  StockReservationNotActiveError,
-} from "@/product/domain/exceptions/product.exceptions";
+import { StockReservationNotActiveError } from "@/product/domain/exceptions/product.exceptions";
+import { StockReservationNotFoundError } from "@/product/application/product.application.exceptions";
 import { StockReservation } from "@/product/domain/entities/stock-reservation.entity";
 import { Product } from "@/product/domain/entities/product.entity";
 import { v4 as uuidv4 } from "uuid";
