+++ src/product/application/use-cases/tier-1-in-domain/reserve-stock.use-case.spec.ts
@@ -1,11 +1,11 @@
 import { Test } from "@nestjs/testing";
 import { ReserveStockUseCase } from "./reserve-stock.use-case";
 import {
-  ProductNotFoundError,
   InactiveProductError,
   InsufficientStockError,
   InvalidQuantityError,
 } from "@/product/domain/exceptions/product.exceptions";
+import { ProductNotFoundError } from "@/product/application/product.application.exceptions";
 import { StockReservation } from "@/product/domain/entities/stock-reservation.entity";
 import { Product } from "@/product/domain/entities/product.entity";
 import { v4 as uuidv4 } from "uuid";
