+++ src/wallet/application/use-cases/tier-1-in-domain/charge-points.use-case.spec.ts
@@ -1,9 +1,7 @@
 import { ChargePointsUseCase } from "./charge-points.use-case";
 import { UserBalance } from "@/wallet/domain/entities/user-balance.entity";
-import {
-  InvalidChargeAmountError,
-  UserBalanceNotFoundError,
-} from "@/wallet/domain/exceptions/point.exceptions";
+import { InvalidChargeAmountError } from "@/wallet/domain/exceptions/point.exceptions";
+import { UserBalanceNotFoundError } from "@/wallet/application/wallet.application.exceptions";
 import { v4 as uuidv4 } from "uuid";
 
 jest.mock("@/wallet/infrastructure/persistence/use-balance.repository");
