+++ src/wallet/application/use-cases/tier-1-in-domain/use-points.use-case.spec.ts
@@ -1,9 +1,7 @@
 import { UsePointsUseCase } from "./use-points.use-case";
 import { UserBalance } from "@/wallet/domain/entities/user-balance.entity";
-import {
-  InsufficientPointBalanceError,
-  UserBalanceNotFoundError,
-} from "@/wallet/domain/exceptions/point.exceptions";
+import { InsufficientPointBalanceError } from "@/wallet/domain/exceptions/point.exceptions";
+import { UserBalanceNotFoundError } from "@/wallet/application/wallet.application.exceptions";
 import { v4 as uuidv4 } from "uuid";
 import { ValidatePointTransactionService } from "@/wallet/domain/services/validate-point-transaction.service";
